+++ src/material/model_stress_strain.rs
